requirements.txt
# Automated CV Submissions
python-dotenv>=1.0.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
